requirements.txt
discord-rewrite==1.0.0
asyncio
aiohttp<2.3.0,>=2.0.0
websockets<5.0,>=4.0s<5.0,>=4.0
async-timeout
requests
